requirements.txt
fastapi==0.75.1
pydantic==1.8.2
passlib==1.7.4
email-validator==1.2.1
uvicorn==0.17.6
python-multipart==0.0.5
opencv-python==4.2.0.34
Pillow==9.1.1
tensorflow==2.7.0
protobuf==3.20.0
nltk==3.7
beautifulsoup4==4.11.1
lxml==4.8.0
langdetect==1.0.9
googletrans==4.0.0rc1
spacy==3.3.0
aiofiles==0.8.0
idna==2.8
h11==0.9.0
